feat(#172): move api and spec
