+++ internal/service/cache/repositorycache.go
@@ -3,7 +3,7 @@ package cache
 import (
 	"context"
 	"fmt"
-	openapi "github.com/Interhyp/metadata-service/api/v1"
+	"github.com/Interhyp/metadata-service/api"
 	"github.com/StephanHCB/go-backend-service-common/api/apierrors"
 )
