+++ internal/service/mapper/generics.go
@@ -3,7 +3,7 @@ package mapper
 import (
 	"context"
 	"fmt"
-	openapi "github.com/Interhyp/metadata-service/api/v1"
+	"github.com/Interhyp/metadata-service/api"
 	"github.com/Interhyp/metadata-service/internal/acorn/errors/nochangeserror"
 	"github.com/StephanHCB/go-backend-service-common/web/middleware/requestid"
 	"github.com/rs/zerolog/log"
