+++ internal/service/owners/owners.go
@@ -3,10 +3,10 @@ package owners
 import (
 	"context"
 	"fmt"
+	"github.com/Interhyp/metadata-service/api"
 	"github.com/Interhyp/metadata-service/internal/acorn/service"
 	"strings"
 
-	openapi "github.com/Interhyp/metadata-service/api/v1"
 	librepo "github.com/StephanHCB/go-backend-service-common/acorns/repository"
 	"github.com/StephanHCB/go-backend-service-common/api/apierrors"
 )
