+++ internal/service/updater/services.go
@@ -3,7 +3,7 @@ package updater
 import (
 	"context"
 	"errors"
-	openapi "github.com/Interhyp/metadata-service/api/v1"
+	"github.com/Interhyp/metadata-service/api"
 	"github.com/Interhyp/metadata-service/internal/acorn/errors/nochangeserror"
 	"github.com/Interhyp/metadata-service/internal/acorn/repository"
 )
