+++ internal/web/util/paramhelper.go
@@ -3,7 +3,7 @@ package util
 import (
 	"context"
 	"encoding/json"
-	openapi "github.com/Interhyp/metadata-service/api/v1"
+	"github.com/Interhyp/metadata-service/api"
 	aulogging "github.com/StephanHCB/go-autumn-logging"
 	"github.com/StephanHCB/go-backend-service-common/api/apierrors"
 	"github.com/go-chi/chi/v5"
